pyproject.toml
[project]
name = "wangame-eval"
version = "0.1.0"
description = "Evaluation pipeline for the WanGame 1.3B Minecraft World Model"
requires-python = ">=3.10"
dependencies = [
    "wandb>=0.16",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
